// backend/src/permissions/interfaces.ts
export enum Action {
  CREATE = 'create',
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete'
}

export enum Resource {
  SOURCE_OF_FUNDS = 'source_of_funds',
  TRANSACTIONS = 'transactions',
  TRANSFERS = 'transfers',
  INVITATIONS = 'invitations',
  RECIPIENTS = 'recipients',
  CATEGORIES = 'categories',
  MEMBERS = 'members',
  PAYMENT_LINKS = 'payment_links',
  FINANCIAL_TRANSACTIONS = 'financial_transactions',
  WALLETS = 'wallets',
  WALLET_GROUPS = 'wallet_groups',
  ASSETS = 'assets',
  CRYPTOCURRENCIES = 'cryptocurrencies',
  SETTINGS = 'settings'
}